RF: Started modularizing iohub eye tracker calibration graphics
RF: Created BaseCalibrationProcedure class, which MouseGaze now uses / extends.
Not yet implemented, but GazePoint and Tobii will also use BaseCalibrationProcedure.
